imap: synchronize a fully fetched message w/ local db
The messages list is first retrieved from Proton via the /messages call.
A JSON representation of each of these messages is saved in a local db.
However, the retrieved JSON description doesn't include the MIMEType
field which is left empty.

This CL fixes this by always sync'ing the local db with the full message
description as soon as it's retrieved from Proton.

Fixes emersion/hydroxide#112.