Replace dependency on boltdb with bbolt v1.4.0
Started failing (notably, on OpenBSD)[1] due to an attempt of a direct
syscall to madvise(2), which on modern OpenBSDs is reachable via a
trampoline in libc.

The boltdb project has been sunset and archived, as the project author
deems the project to have achieved it's goals and is thus complete.[2]

bbolt is a CoreOS fork of boltdb, with additional features in addition
to fixing the madvise(2) call issue by relying on the `unix` module
instead.[3]

v1.4.0 is the latest release at the time of writing, and after
downloading ~4000 mail messages across multiple mailboxes, and managing
multiple user accounts, has worked without issue.

[1]: https://github.com/emersion/hydroxide/issues/284
[2]: https://github.com/boltdb/bolt?tab=readme-ov-file#a-message-from-the-author
[3]: https://github.com/etcd-io/bbolt/blob/0d516858c4bad9ad7c7ccb637adc32bcd78ca8cb/bolt_unix.go#L62